refactor: move countingIterator to separate file
Extract countingIterator from archive.go into its own file for better
code organization. The main archive.go file should focus on core LSM-tree
operations rather than iterator utilities.
